refactor(ui): move SafeSnap strategy discovery out of useTreasuries
SafeSnap config parsing and the avatar() resolution had grown to dominate
useTreasuries, a network-agnostic composable shared by evm, starknet and
offchain. Move that read-side discovery to helpers/safesnap/strategies.ts,
next to the transaction parser and the plugin builder, so all SafeSnap
knowledge sits together and useTreasuries goes back to mapping treasuries.
The extracted getSafeSnapStrategies gets unit tests for the config
validation (placeholder rejection, non-numeric network, non-array safes),
which the composable could not exercise in isolation.
